backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.31.1
python-dotenv==1.0.1
boto3==1.35.57
